[1.4] Ported example tiles and patches (#1462)
* + Restore hooks for tileDrawing
+ Porting ExampleAnimatedTile to 1.4
* fix ExampleTorch

* + Restore hooks for tileDrawing
+ Porting ExampleAnimatedTile to 1.4
* fix ExampleTorch

* * Fixes in ExampleAnimatedTile
+ Rework VoidMonolith into ExampleAnimatedGlowmaskTile

* Rediff

* Moved TileLoader.Pre/PostDraw implementation.

* Relocated SetAnimationFrame

* A small TODO note.
