+++ ExampleMod/Content/Tiles/ExampleAnimatedGlowmaskTile.png
