+++ ExampleMod/Content/Tiles/ExampleAnimatedGlowmaskTileItem.png
