+++ ExampleMod/Content/Tiles/ExampleAnimatedGlowmaskTile_Glow.png
