+++ ExampleMod/Content/Tiles/ExampleAnimatedTile.png
