+++ ExampleMod/Content/Tiles/ExampleAnimatedTileItem.png
